Changed the abbreviated state name (SA) to the full name (South Australia) to be consistent with other holiday providers.
