+++ src/Yasumi/Provider/Australia/SouthAustralia.php
@@ -23,7 +23,7 @@ use Yasumi\Provider\DateTimeZoneFactory;
  * Provider for all holidays in South Australia (Australia).
  *
  */
-class SA extends Australia
+class SouthAustralia extends Australia
 {
     /**
      * Code to identify this Holiday Provider. Typically this is the ISO3166 code corresponding to the respective
